backend/requirements.txt
fastapi
uvicorn
pillow
numpy
torch
torchvision
faiss-cpu
python-multipart
